protection + delete series miroir.

+++ picsou/picsou/src/main/java/org/designup/picsou/triggers/projects/ProjectItemToSeriesTrigger.java
@@ -75,8 +75,20 @@ public class ProjectItemToSeriesTrigger implements ChangeSetListener {
           !repository.contains(Transaction.TYPE,
                                and(fieldEquals(Transaction.SERIES, seriesId),
                                    isFalse(Transaction.PLANNED)))) {
-        repository.update(item.getKey(), ProjectItem.SERIES, null);
-        repository.delete(seriesKey);
+        Glob series = repository.get(seriesKey);
+        if (series.get(Series.MIRROR_SERIES) != null){
+          if (!repository.contains(Transaction.TYPE,
+                              and(fieldEquals(Transaction.SERIES, seriesId),
+                                  isFalse(Transaction.PLANNED)))) {
+            repository.update(item.getKey(), ProjectItem.SERIES, null);
+            repository.delete(seriesKey);
+            repository.delete(KeyBuilder.newKey(Series.TYPE, series.get(Series.ID)));
+          }
+        }
+        else {
+          repository.update(item.getKey(), ProjectItem.SERIES, null);
+          repository.delete(seriesKey);
+        }
       }
     }
   }
